+++ product-groups.php
@@ -2,8 +2,8 @@
 /**
  * Plugin Name: Product Groups (Link + Affiliate)
  * Plugin URI: https://github.com/alireza-kh95/product-groups
- * Description: Create and display responsive groups of Digikala products using product and affiliate links with automatic API caching and GitHub auto-updates.
- * Version: 1.3.2
+ * Description: Create and display responsive groups of Digikala & Snapp Shop products using product and affiliate links with automatic API caching and GitHub auto-updates.
+ * Version: 1.4.0
  * Author: Alireza Khosravani
  * Author URI: https://github.com/alireza-kh95
  * License: GPL-2.0-or-later
@@ -21,7 +21,7 @@ if ( ! defined( 'ABSPATH' ) ) {
 }
 
 // Define Plugin Constants
-define( 'PRODUCT_GROUPS_VERSION', '1.3.2' );
+define( 'PRODUCT_GROUPS_VERSION', '1.4.0' );
 define( 'PRODUCT_GROUPS_FILE', __FILE__ );
 define( 'PRODUCT_GROUPS_BASENAME', plugin_basename( __FILE__ ) );
 define( 'PRODUCT_GROUPS_DIR', plugin_dir_path( __FILE__ ) );